requirements.txt
numpy>=1.22.0
opencv-python>=4.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
torch>=2.0.0
scipy>=1.8.0
scikit-learn>=1.0.0
tqdm>=4.60.0
pillow>=9.0.0
requests>=2.25.0
timm>=0.9.0
diffusers>=0.20.0
transformers>=4.30.0
segment-anything>=1.0
hdbscan>=0.8.29
llama-cpp-python>=0.2.0
pandas>=1.3.0
tabulate>=0.8.0